requirements.txt
paho-mqtt==1.5.1
python-decouple==3.4
redis==3.5.3
